# internal/api/assets/openapi.yaml
openapi: 3.0.3
info:
  title: Baseline API
  version: "1.0.0"
  description: "Baseline policy and dashboard API contract for implemented /v1 endpoints. Fields marked with x-baseline-implementation=reserved_noop are accepted for forward compatibility but not currently enforced."
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /v1/auth/session:
    post:
      summary: Create dashboard session cookie
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionState"
        "403":
          $ref: "#/components/responses/ErrorResponse"
    get:
      summary: Get current dashboard session
      responses:
        "200":
          description: Session active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionState"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    delete:
      summary: End dashboard session
      parameters:
        - in: header
          name: X-Baseline-CSRF
          required: true
          schema:
            type: string
            enum: ["1"]
      responses:
        "200":
          description: Session terminated
          content:
            application/json:
              schema:
                type: object
                properties:
                  active:
                    type: boolean
                    example: false
  /v1/auth/reauth:
    post:
      summary: Issue short-lived re-auth token for sensitive actions
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: header
          name: X-Baseline-CSRF
          required: false
          schema:
            type: string
            enum: ["1"]
      responses:
        "200":
          description: Re-auth token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  reauth_token:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ErrorResponse"
  /v1/auth/register:
    post:
      summary: Register API key via enrollment token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enrollment_token]
              properties:
                enrollment_token:
                  type: string
      responses:
        "201":
          description: API key issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  role:
                    $ref: "#/components/schemas/Role"
                  api_key:
                    type: string
                  auth_mode:
                    type: string
                    example: api_key
        "403":
          $ref: "#/components/responses/ErrorResponse"
  /v1/api-keys:
    get:
      summary: List API keys (metadata only)
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: API key inventory
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_keys:
                    type: array
                    items:
                      $ref: "#/components/schemas/APIKeyMetadata"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      summary: Issue API key (admin)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: header
          name: X-Baseline-CSRF
          required: false
          schema:
            type: string
            enum: ["1"]
        - in: header
          name: Idempotency-Key
          required: false
          description: Reserved for future idempotent API key issuance. Currently accepted but not enforced.
          x-baseline-implementation: reserved_noop
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAPIKeyRequest"
      responses:
        "201":
          description: API key created (secret returned once)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIKeyMetadata"
                  - type: object
                    properties:
                      api_key:
                        type: string
        "403":
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/me/api-keys:
    get:
      summary: List API keys owned by the authenticated user
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: query
          name: include_revoked
          required: false
          schema:
            type: boolean
        - in: query
          name: limit
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: User-owned API key inventory
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_keys:
                    type: array
                    items:
                      $ref: "#/components/schemas/APIKeyMetadata"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ErrorResponse"
    post:
      summary: Create API key for authenticated user
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: header
          name: X-Baseline-CSRF
          required: false
          schema:
            type: string
            enum: ["1"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAPIKeyRequest"
      responses:
        "201":
          description: API key created (secret returned once)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIKeyMetadata"
                  - type: object
                    properties:
                      api_key:
                        type: string
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/users:
    get:
      summary: List users (admin)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: query
          name: role
          required: false
          schema:
            $ref: "#/components/schemas/Role"
        - in: query
          name: status
          required: false
          schema:
            $ref: "#/components/schemas/UserStatus"
        - in: query
          name: q
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
        - in: query
          name: offset
          required: false
          schema:
            type: integer
        - in: query
          name: sort_by
          required: false
          schema:
            type: string
            enum: [user, role, status, last_login_at, created_at, updated_at]
        - in: query
          name: sort_dir
          required: false
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: User inventory
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/UserRecord"
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
                  has_more:
                    type: boolean
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/integrations/github/webhook:
    post:
      summary: Receive GitHub webhook events (signed)
      parameters:
        - in: header
          name: X-Hub-Signature-256
          required: true
          schema:
            type: string
        - in: header
          name: X-GitHub-Event
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "202":
          description: Webhook accepted
        "403":
          $ref: "#/components/responses/ErrorResponse"
  /v1/integrations/gitlab/webhook:
    post:
      summary: Receive GitLab webhook events (token)
      parameters:
        - in: header
          name: X-Gitlab-Token
          required: true
          schema:
            type: string
        - in: header
          name: X-Gitlab-Event
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "202":
          description: Webhook accepted
        "403":
          $ref: "#/components/responses/ErrorResponse"
  /v1/integrations/github/check-runs:
    post:
      summary: Publish GitHub check run status
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: header
          name: X-Baseline-CSRF
          required: false
          schema:
            type: string
            enum: ["1"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GitHubCheckRunRequest"
      responses:
        "202":
          description: Check run accepted by integration layer
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "502":
          $ref: "#/components/responses/ErrorResponse"
  /v1/integrations/gitlab/statuses:
    post:
      summary: Publish GitLab commit status
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: header
          name: X-Baseline-CSRF
          required: false
          schema:
            type: string
            enum: ["1"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GitLabStatusRequest"
      responses:
        "202":
          description: Commit status accepted by integration layer
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "502":
          $ref: "#/components/responses/ErrorResponse"
  /v1/integrations/jobs:
    get:
      summary: List integration jobs (queue/retry/backoff visibility)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: provider
          required: false
          schema:
            type: string
            enum: [github, gitlab]
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [pending, running, succeeded, failed]
      responses:
        "200":
          description: Integration jobs list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationJobsResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /v1/integrations/secrets:
    post:
      summary: Update runtime integration secrets/config (admin only)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: header
          name: X-Baseline-CSRF
          required: false
          schema:
            type: string
            enum: ["1"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegrationSecretsUpdateRequest"
      responses:
        "200":
          description: Integration secret references updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: array
                    items:
                      type: string
                  count:
                    type: integer
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
  /v1/projects:
    get:
      summary: List projects
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
    post:
      summary: Create project
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: header
          name: X-Baseline-CSRF
          required: false
          schema:
            type: string
            enum: ["1"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/scans:
    get:
      summary: List scans
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: query
          name: project_id
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Scan list
          content:
            application/json:
              schema:
                type: object
                properties:
                  scans:
                    type: array
                    items:
                      $ref: "#/components/schemas/ScanSummary"
    post:
      summary: Ingest scan
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: header
          name: X-Baseline-CSRF
          required: false
          schema:
            type: string
            enum: ["1"]
        - in: header
          name: Idempotency-Key
          required: false
          description: Idempotency key for safe scan ingestion retries.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScanRequest"
      responses:
        "201":
          description: Scan accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanSummary"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/dashboard:
    get:
      summary: Dashboard aggregate summary
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Dashboard aggregates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /v1/dashboard/capabilities:
    get:
      summary: Dashboard role/capability flags
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Capabilities for current principal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardCapabilitiesResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /v1/dashboard/activity:
    get:
      summary: Dashboard activity feed
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: cursor
          required: false
          schema:
            type: string
        - in: query
          name: type
          required: false
          schema:
            type: string
            enum: [scan, audit, integration]
        - in: query
          name: actor
          required: false
          schema:
            type: string
        - in: query
          name: from
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: false
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Activity feed page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardActivityResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /v1/policies:
    get:
      summary: List policy summaries
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Policies
...
  /v1/rulesets:
    post:
      summary: Publish ruleset (admin)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: header
          name: X-Baseline-CSRF
          required: false
          schema:
            type: string
            enum: ["1"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRulesetRequest"
      responses:
        "201":
          description: Ruleset created
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/audit/events:
    get:
      summary: List audit events
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: query
          name: project_id
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Event stream
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: APIKey
    cookieAuth:
      type: apiKey
      in: cookie
      name: baseline_dashboard_session
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnauthorizedError:
      description: Unauthorized
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    Role:
      type: string
      enum: [viewer, operator, admin]
    UserStatus:
      type: string
      enum: [active, suspended]
    SessionState:
      type: object
      properties:
        user:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        auth_mode:
          type: string
        active:
          type: boolean
    CreateAPIKeyRequest:
      type: object
      properties:
        name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
    APIKeyMetadata:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        prefix:
          type: string
        source:
          type: string
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
        created_by_user_id:
          type: string
        owner_user_id:
          type: string
        owner_subject:
          type: string
        owner_email:
          type: string
        revoked:
          type: boolean
        revoked_at:
          type: string
          format: date-time
        revoked_by_user_id:
          type: string
        revocation_reason:
          type: string
    UserRecord:
      type: object
      properties:
        id:
          type: string
        display_name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        status:
          $ref: "#/components/schemas/UserStatus"
        last_login_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    GitHubCheckRunRequest:
      type: object
      required: [owner, repository, head_sha, name]
      properties:
        owner:
          type: string
        repository:
          type: string
        head_sha:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [queued, in_progress, completed]
        conclusion:
          type: string
        details_url:
          type: string
        external_id:
          type: string
        output:
          type: object
          properties:
            title:
              type: string
            summary:
              type: string
    GitLabStatusRequest:
      type: object
      required: [project_id, sha, state]
      properties:
        project_id:
          type: string
        sha:
          type: string
        state:
          type: string
          enum: [pending, running, success, failed, canceled, skipped]
        name:
          type: string
        target_url:
          type: string
        description:
          type: string
        ref:
          type: string
    IntegrationJobSummary:
      type: object
      properties:
        id:
          type: string
        provider:
          type: string
        job_type:
          type: string
        project_ref:
          type: string
        external_ref:
          type: string
        status:
          type: string
          enum: [pending, running, succeeded, failed]
        attempt_count:
          type: integer
        max_attempts:
          type: integer
        last_error:
          type: string
        next_attempt_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    IntegrationJobsResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/IntegrationJobSummary"
    IntegrationSecretsUpdateRequest:
      type: object
      properties:
        github_webhook_secret:
          type: string
        gitlab_webhook_token:
          type: string
        github_api_token:
          type: string
        gitlab_api_token:
          type: string
        github_api_url:
          type: string
        gitlab_api_url:
          type: string
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        repository_url:
          type: string
        default_branch:
          type: string
        policy_set:
          type: string
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        id:
          type: string
          description: Optional caller-provided identifier. If omitted, server generates one.
        name:
          type: string
        repository_url:
          type: string
          description: Optional repository metadata field used for display and audit context.
        default_branch:
          type: string
          description: Optional project default branch. Defaults to main.
        policy_set:
          type: string
          description: Optional policy set label. Defaults to baseline:prod.
    ScanViolation:
      type: object
      properties:
        policy_id:
          type: string
        severity:
          type: string
        message:
          type: string
    ScanSummary:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        commit_sha:
          type: string
        status:
          type: string
        violations:
          type: array
          items:
            $ref: "#/components/schemas/ScanViolation"
        created_at:
          type: string
          format: date-time
    CreateScanRequest:
      type: object
      required: [project_id]
      properties:
        id:
          type: string
          description: Optional caller-provided scan id. If omitted, server generates one.
        project_id:
          type: string
        commit_sha:
          type: string
        status:
          type: string
        violations:
          type: array
          items:
            $ref: "#/components/schemas/ScanViolation"
...
    CreateRulesetRequest:
      type: object
      required: [policy_names]
      properties:
        version:
          type: string
        description:
          type: string
        policy_names:
          type: array
          items:
            type: string
    DashboardMetrics:
      type: object
      properties:
        projects:
          type: integer
        scans:
          type: integer
        failing_scans:
          type: integer
        blocking_violations:
          type: integer
    DashboardViolationCount:
      type: object
      properties:
        policy_id:
          type: string
        count:
          type: integer
    AuditEvent:
      type: object
      properties:
        event_type:
          type: string
        project_id:
          type: string
        scan_id:
          type: string
        actor:
          type: string
        request_id:
          type: string
        created_at:
          type: string
          format: date-time
    DashboardSummary:
      type: object
      properties:
        metrics:
          $ref: "#/components/schemas/DashboardMetrics"
        recent_scans:
          type: array
          items:
            $ref: "#/components/schemas/ScanSummary"
        top_violations:
          type: array
          items:
            $ref: "#/components/schemas/DashboardViolationCount"
        recent_events:
          type: array
          items:
            $ref: "#/components/schemas/AuditEvent"
    DashboardCapabilities:
      type: object
      properties:
        dashboard.view:
          type: boolean
        projects.read:
          type: boolean
        projects.write:
          type: boolean
        scans.read:
          type: boolean
        scans.run:
          type: boolean
        api_keys.read:
          type: boolean
        api_keys.write:
          type: boolean
        audit.read:
          type: boolean
        integrations.read:
          type: boolean
        integrations.write:
          type: boolean
        integrations.secrets.write:
          type: boolean
    DashboardCapabilitiesResponse:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/Role"
        source:
          type: string
          enum: [api_key, session]
        capabilities:
          $ref: "#/components/schemas/DashboardCapabilities"
...
    DashboardActivityResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DashboardActivityItem"
        next_cursor:
          type: string